# oceanembed-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OceanEmbed API",
    "description": "Subsurface Ocean Temperature Reconstruction Backend (SIH26066)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/argo/validation": {
      "get": {
        "tags": [
          "Argo Validation & Profiles"
        ],
        "summary": "Get Argo matchup validation statistics",
        "operationId": "get_argo_validation_api_v1_argo_validation_get",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Start date (YYYY-MM-DD)",
              "examples": [
                "2024-01-01"
              ],
              "default": "2024-01-01",
              "title": "Date From"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "End date (YYYY-MM-DD)",
              "examples": [
                "2026-06-30"
              ],
              "default": "2026-06-30",
              "title": "Date To"
            },
            "description": "End date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArgoValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/argo/profiles": {
      "get": {
        "tags": [
          "Argo Validation & Profiles"
        ],
        "summary": "Get nearby Argo float profiles",
        "operationId": "get_argo_profiles_api_v1_argo_profiles_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Center latitude \u00b0N",
              "examples": [
                12.5
              ],
              "title": "Lat"
            },
            "description": "Center latitude \u00b0N"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Center longitude \u00b0E",
              "examples": [
                65.0
              ],
              "title": "Lon"
            },
            "description": "Center longitude \u00b0E"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1000.0,
              "minimum": 1.0,
              "description": "Radius in km",
              "examples": [
                100.0
              ],
              "default": 100.0,
              "title": "Radius Km"
            },
            "description": "Radius in km"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target date (YYYY-MM-DD)",
              "examples": [
                "2025-05-15"
              ],
              "title": "Date"
            },
            "description": "Target date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArgoProfilesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/ohc": {
      "get": {
        "tags": [
          "Derived Ocean Products"
        ],
        "summary": "Get Ocean Heat Content (OHC)",
        "operationId": "get_ohc_api_v1_products_ohc_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude \u00b0N",
              "examples": [
                12.5
              ],
              "title": "Lat"
            },
            "description": "Latitude \u00b0N"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude \u00b0E",
              "examples": [
                65.0
              ],
              "title": "Lon"
            },
            "description": "Longitude \u00b0E"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target date (YYYY-MM-DD)",
              "examples": [
                "2025-05-15"
              ],
              "title": "Date"
            },
            "description": "Target date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OHCProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/marine-heatwave": {
      "get": {
        "tags": [
          "Derived Ocean Products"
        ],
        "summary": "Get Marine Heatwave (MHW) assessment",
        "operationId": "get_marine_heatwave_api_v1_products_marine_heatwave_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude \u00b0N",
              "examples": [
                12.5
              ],
              "title": "Lat"
            },
            "description": "Latitude \u00b0N"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude \u00b0E",
              "examples": [
                65.0
              ],
              "title": "Lon"
            },
            "description": "Longitude \u00b0E"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target date (YYYY-MM-DD)",
              "examples": [
                "2025-05-15"
              ],
              "title": "Date"
            },
            "description": "Target date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarineHeatwaveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/cyclone-risk": {
      "get": {
        "tags": [
          "Derived Ocean Products"
        ],
        "summary": "Get Tropical Cyclone Heat Potential & Risk",
        "operationId": "get_cyclone_risk_api_v1_products_cyclone_risk_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude \u00b0N",
              "examples": [
                12.5
              ],
              "title": "Lat"
            },
            "description": "Latitude \u00b0N"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude \u00b0E",
              "examples": [
                65.0
              ],
              "title": "Lon"
            },
            "description": "Longitude \u00b0E"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target date (YYYY-MM-DD)",
              "examples": [
                "2025-05-15"
              ],
              "title": "Date"
            },
            "description": "Target date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycloneRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ingest/trigger": {
      "post": {
        "tags": [
          "Admin & Pipeline Control"
        ],
        "summary": "Trigger ingestion pipeline job",
        "operationId": "trigger_ingestion_api_v1_admin_ingest_trigger_post",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Target date YYYY-MM-DD",
              "examples": [
                "2025-05-15"
              ],
              "default": "2025-05-15",
              "title": "Date"
            },
            "description": "Target date YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ingest/status/{job_id}": {
      "get": {
        "tags": [
          "Admin & Pipeline Control"
        ],
        "summary": "Get ingestion job status",
        "operationId": "get_ingestion_status_api_v1_admin_ingest_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArgoFloatProfile": {
        "properties": {
          "float_id": {
            "type": "string",
            "title": "Float Id",
            "examples": [
              "ARGO_2901482"
            ]
          },
          "lat": {
            "type": "number",
            "title": "Lat",
            "examples": [
              12.6
            ]
          },
          "lon": {
            "type": "number",
            "title": "Lon",
            "examples": [
              80.1
            ]
          },
          "date": {
            "type": "string",
            "title": "Date",
            "examples": [
              "2025-05-14"
            ]
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km",
            "examples": [
              15.2
            ]
          },
          "profile": {
            "items": {
              "$ref": "#/components/schemas/ProfilePoint"
            },
            "type": "array",
            "title": "Profile"
          }
        },
        "type": "object",
        "required": [
          "float_id",
          "lat",
          "lon",
          "date",
          "distance_km",
          "profile"
        ],
        "title": "ArgoFloatProfile"
      },
      "ArgoProfilesResponse": {
        "properties": {
          "requested_location": {
            "$ref": "#/components/schemas/LocationPoint"
          },
          "radius_km": {
            "type": "number",
            "title": "Radius Km",
            "examples": [
              100.0
            ]
          },
          "date": {
            "type": "string",
            "title": "Date",
            "examples": [
              "2025-05-15"
            ]
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ArgoFloatProfile"
            },
            "type": "array",
            "title": "Profiles"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "mock",
            "examples": [
              "mock"
            ]
          }
        },
        "type": "object",
        "required": [
          "requested_location",
          "radius_km",
          "date",
          "profiles"
        ],
        "title": "ArgoProfilesResponse"
      },
      "ArgoValidationResponse": {
        "properties": {
          "date_from": {
            "type": "string",
            "title": "Date From",
            "examples": [
              "2024-01-01"
            ]
          },
          "date_to": {
            "type": "string",
            "title": "Date To",
            "examples": [
              "2026-06-30"
            ]
          },
          "n_matchups": {
            "type": "integer",
            "title": "N Matchups",
            "examples": [
              1450
            ]
          },
          "overall": {
            "$ref": "#/components/schemas/OverallValidation"
          },
          "by_depth": {
            "items": {
              "$ref": "#/components/schemas/DepthValidation"
            },
            "type": "array",
            "title": "By Depth"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "mock",
            "examples": [
              "mock"
            ]
          }
        },
        "type": "object",
        "required": [
          "date_from",
          "date_to",
          "n_matchups",
          "overall",
          "by_depth"
        ],
        "title": "ArgoValidationResponse"
      },
      "CycloneRiskResponse": {
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationPoint"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "examples": [
              "2025-05-15"
            ]
          },
          "tchp_score_kj_cm2": {
            "type": "number",
            "title": "Tchp Score Kj Cm2",
            "description": "Tropical Cyclone Heat Potential in kJ/cm^2",
            "examples": [
              62.4
            ]
          },
          "isotherm_26c_depth_m": {
            "type": "number",
            "title": "Isotherm 26C Depth M",
            "description": "Depth of 26\u00b0C isotherm in meters",
            "examples": [
              65.0
            ]
          },
          "risk_category": {
            "type": "string",
            "title": "Risk Category",
            "description": "Low, Moderate, or High",
            "examples": [
              "Moderate"
            ]
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "mock",
            "examples": [
              "mock"
            ]
          }
        },
        "type": "object",
        "required": [
          "location",
          "date",
          "tchp_score_kj_cm2",
          "isotherm_26c_depth_m",
          "risk_category"
        ],
        "title": "CycloneRiskResponse"
      },
      "DepthRange": {
        "properties": {
          "min": {
            "type": "number",
            "title": "Min",
            "default": 0.0,
            "examples": [
              0.0
            ]
          },
          "max": {
            "type": "number",
            "title": "Max",
            "default": 450.0,
            "examples": [
              450.0
            ]
          }
        },
        "type": "object",
        "title": "DepthRange"
      },
      "DepthValidation": {
        "properties": {
          "depth_m": {
            "type": "number",
            "title": "Depth M",
            "examples": [
              0.0
            ]
          },
          "rmse_c": {
            "type": "number",
            "title": "Rmse C",
            "examples": [
              0.25
            ]
          },
          "correlation": {
            "type": "number",
            "title": "Correlation",
            "examples": [
              0.98
            ]
          }
        },
        "type": "object",
        "required": [
          "depth_m",
          "rmse_c",
          "correlation"
        ],
        "title": "DepthValidation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "examples": [
              "JOB_a1b2c3d4"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "succeeded"
            ]
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "mock",
            "examples": [
              "mock"
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "IngestStatusResponse"
      },
      "IngestTriggerResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "examples": [
              "JOB_a1b2c3d4"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "running",
            "examples": [
              "running"
            ]
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "mock",
            "examples": [
              "mock"
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "started_at"
        ],
        "title": "IngestTriggerResponse"
      },
      "LocationPoint": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat",
            "examples": [
              12.5
            ]
          },
          "lon": {
            "type": "number",
            "title": "Lon",
            "examples": [
              80.0
            ]
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "LocationPoint"
      },
      "MarineHeatwaveResponse": {
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationPoint"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "examples": [
              "2025-05-15"
            ]
          },
          "is_heatwave": {
            "type": "boolean",
            "title": "Is Heatwave",
            "examples": [
              false
            ]
          },
          "sst_c": {
            "type": "number",
            "title": "Sst C",
            "examples": [
              28.4
            ]
          },
          "climatological_threshold_c": {
            "type": "number",
            "title": "Climatological Threshold C",
            "default": 29.5,
            "examples": [
              29.5
            ]
          },
          "severity_category": {
            "type": "string",
            "title": "Severity Category",
            "description": "None, Moderate, Strong, Severe, or Extreme",
            "examples": [
              "None"
            ]
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "mock",
            "examples": [
              "mock"
            ]
          }
        },
        "type": "object",
        "required": [
          "location",
          "date",
          "is_heatwave",
          "sst_c",
          "severity_category"
        ],
        "title": "MarineHeatwaveResponse"
      },
...
      "OHCProductResponse": {
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationPoint"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "examples": [
              "2025-05-15"
            ]
          },
          "ohc_value": {
            "type": "number",
            "title": "Ohc Value",
            "description": "Ocean Heat Content value",
            "examples": [
              3.68
            ]
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "GJ/m^2",
            "examples": [
              "GJ/m^2"
            ]
          },
          "ohc_joules_m2": {
            "type": "number",
            "title": "Ohc Joules M2",
            "description": "Ocean Heat Content in Joules per square meter",
            "examples": [
              3680000000.0
            ]
          },
          "reference_temperature_c": {
            "type": "number",
            "title": "Reference Temperature C",
            "default": 0.0,
            "examples": [
              0.0
            ]
          },
          "depth_range_m": {
            "$ref": "#/components/schemas/DepthRange"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "mock",
            "examples": [
              "mock"
            ]
          }
        },
        "type": "object",
        "required": [
          "location",
          "date",
          "ohc_value",
          "ohc_joules_m2",
          "depth_range_m"
        ],
        "title": "OHCProductResponse"
      },
      "OverallValidation": {
        "properties": {
          "rmse_c": {
            "type": "number",
            "title": "Rmse C",
            "examples": [
              0.42
            ]
          },
          "correlation": {
            "type": "number",
            "title": "Correlation",
            "examples": [
              0.96
            ]
          }
        },
        "type": "object",
        "required": [
          "rmse_c",
          "correlation"
        ],
        "title": "OverallValidation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}